Clean up callback for prediction writer
